% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_rawdata.R
\name{compare_rawdata}
\alias{compare_rawdata}
\title{compare rawdata}
\usage{
compare_rawdata(...)
}
\arguments{
\item{...}{named list of rawdata files}
}
\value{
data.frame
}
\description{
Given multiple rawdata files, compares the basecalls at each position and
returns a data.frame of differences.
}
