% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsid_2_genes.R
\name{load_rsid}
\alias{load_rsid}
\title{load rsid}
\usage{
load_rsid(rsid_file = system.file("extdata/rsid.rds", package =
  "pkg23andMeGenome"))
}
\arguments{
\item{rsid_file}{the file to use to load}
}
\value{
list
}
\description{
Loads in the RSID file and names the entries so they can be accessed.
}
