% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rsid.R
\name{query_rsidlist}
\alias{query_rsidlist}
\title{get a bunch of rsid}
\usage{
query_rsidlist(rsid_list, n_query = 250, start = 1)
}
\arguments{
\item{rsid_list}{the list of rsids to query}

\item{n_query}{how many to query at each time (default = 100)}

\item{start}{which item in the list to start with (allows for restarts)}
}
\description{
given a list of rsid, politely get the data from entrez
}
