% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rsid.R
\name{read_rsid}
\alias{read_rsid}
\title{read rsid file}
\usage{
read_rsid(rsid_file)
}
\arguments{
\item{rsid_file}{the file to read}
}
\description{
given a genome file, read it into a data.frame that can be queried and compared
with others.
}
