% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rsid.R
\name{read_rsid_files}
\alias{read_rsid_files}
\title{read rsid files}
\usage{
read_rsid_files(rsid_files, save_file)
}
\arguments{
\item{rsid_files}{the set of files to read in}

\item{save_file}{where to save the full set}
}
\description{
Reads in a bunch of RSID queries from a set of files and saves it to a single
object that can be queried.
}
