% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsid_2_genes.R
\name{rsid_2_vars}
\alias{rsid_2_vars}
\title{table of rsid to variables}
\usage{
rsid_2_vars(rsids, ...)
}
\arguments{
\item{rsids}{the list of named rsids to use}

\item{...}{the variables (quoted) to map}
}
\value{
data.frame
}
\description{
Create a table of RSIDs to selected variables
}
